DEVDOCS-97 moving image files and corrections from Cory
Implemented Corrections from Cory regarding installation process and image formatting issues.

This led me to a preparatory step for reorganizing image files so that they can be referenced with absolute paths in dedicated image folders instead of spread out it image sub-folders (which right now is leading to several duplicates of the same file).